+++ Editor/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 using System.Runtime.CompilerServices;
 using Unity.Cloud.Common;
 
-[assembly: ApiSourceVersion("com.unity.cloud.common.editor", "1.1.3")]
+[assembly: ApiSourceVersion("com.unity.cloud.common.editor", "1.1.4")]
 [assembly: InternalsVisibleTo("Unity.Cloud.Identity.Editor")]
 [assembly: InternalsVisibleTo("Unity.Cloud.Identity.Tests.Editor")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]